feat(docs-worklets): Setup `Algolia` for worklets docs (#7762)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR adds support for `Algolia` docs search, which is needed to be
able to use the search engine in `docusarus`.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
<img width="1488" alt="image"
src="https://github.com/user-attachments/assets/ebce8740-3ce6-4919-baa7-1ed2b4a4967b"
/>

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->