feat(Worklets): RuntimeManager (#7739)
## Summary

Introducing a new class, `RuntimeManager`, responsible for creating new
Worklet Runtimes and storing information about them, but not owning
them. A runtime is still owned by its creator.

I have added `getAllRuntimes` method for the purpose of hot-reload in
Bundle Mode. During hot-reload I need to forward the update to all
runtimes, that's why I need a class that allows that. In the future we
might handle hot-reloads some other way, but this is the way to go now.

- https://github.com/software-mansion/react-native-reanimated/pull/7617
will make the code less redundant.

## Test plan

Run the app, see that spawning a new runtime/UI runtime works properly.
Reload the app.