Update react-native-is-edge-to-edge to 1.2.1 (#7729)
## Summary

This PR updates `react-native-is-edge-to-edge` to the latest version:
`1.2.1`.

As `edgeToEdgeEnabled` will be available in the next React Native
release (https://github.com/facebook/react-native/pull/52088), we need a
new release of `react-native-reanimated`, as it uses a fixed version of
the dependency for the moment.

To prevent this kind of issue, if it must happens again, I removed the
strict requirement on the version number (for reference,
`react-native-screens` doesn't fix the version
https://github.com/software-mansion/react-native-screens/blob/62eb784c6457be366a2daf6d6d074b7a586fa61e/package.json#L72)