refactor: withSpring end conditions (#7803)
## Summary

Previous `withSpring` end conditions were based on two parameters:
`restDisplacementThreshold` - how far the spring is from its resting
point and `restSpeedThreshold` - the speed of the spring. These
parameters were absolute and adjusted for animations in ranges greater
than 0-1. For animations in a lesser range the user would need to
overwrite the `restDisplacementThreshold` and `restSpeedThreshold`
parameters so the animation wouldn't snap immediately after starting.
However, since these parameters were absolute it wasn't so obvious what
the values should've been.

Instead, we replace both these parameters with one `energyThreshold`
which tells at what fraction of the initial energy the animation should
snap. This value is relative, quadratic and bound to both position and
speed of the spring. In my testing it proved much more reliable than the
fixed absolute threshold and didn't need to be overridden in any case.

To make equations make sense I had to fix the coordinates system used in
withSpring animation. Now the resting point is always at relative 0,
which makes things more readable in my opinion. If there's an animation
from 0 to 100, it gets translated to (-100, 0). If it's from 100 to 0,
it gets translated to (100, 0). Thanks to this we don't have to compare
the start with the end to figure out whether we should use a minus sign
somewhere or not.

## Test plan

I added an example that shows that the behavior of the spring works as
expected.

`ChatHeads` example shows that chained `withSpring` animations work too.