[Fabric LA] Add progress batching (#7798)
## Summary
This PR changes the way we apply layout animations. Now we call the
`notifyDelegateOfUpdates` method only once per frame, instead of doing
that for every updated component separately. This improves performance,
fixes animations with `duration=0` and possibly helps with some timing
issues, when the view could be updated before it's mounted (#7493 - but
it's probably not fully solved by this PR).

## Test plan

Checks all LA related examples.

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>