[Fabric LA] Mark deleted nodes as DELETED (#7818)
## Summary

This PR fixes crashes that became apparent after #7798. The issue is
that we would add DELETE mutations for some views multiple times,
because in some code-paths we were not marking deleted views as DELETED.
I think that the issue was not manifesting before #7798 because then we
would handle removals separately for each view.

## Test plan

Spam the `[LA] Entering and Exiting with Layout` example and check for
crashes.