Add `DISABLE_COMMIT_PAUSING_MECHANISM` feature flag (#7852)
## Summary

This PR adds a flag that allows users to disable the commit pausing
mechanism in reanimated. The mechanism is used to avoid starving out the
react renderer, by continuously pushing short reanimated updates. With
[this](https://github.com/facebook/react-native/pull/52645) update
coming to RN, we should be soon able to remove the mechanism altogether.
For now we want to ship it as a opt-in feature-flag.

## Test plan