Restore fast path with `synchronouslyUpdateViewOnUIThread` on Android (#7823)
## Summary

This PR restores a fast path for updating animated styles and props
using `synchronouslyUpdateViewOnUIThread` method on Android, currently
hidden behind a static feature flag.

The optimization was removed in
https://github.com/software-mansion/react-native-reanimated/pull/7014
due to the fact that it broke view measurement and touch detection
system e.g. when using `Pressable` from react-native.

## How to test out this change

If you want to test out the new experimental behavior, please:
1. Make sure you're using `react-native-reanimated@4.0.0` or newer
2. Enable `ANDROID_SYNCHRONOUSLY_UPDATE_UI_PROPS` feature flag in your
app's `package.json`:
```json
"reanimated": {
  "staticFeatureFlags": {
    "ANDROID_SYNCHRONOUSLY_UPDATE_UI_PROPS": true
  }
}
```
3. If you use `Pressable` component from `react-native`, use the one
from `react-native-gesture-handler`. That's because it attaches to the
platform view as opposed to the one from `react-native`.

## Disclaimers

Keep in mind that the fast path works only on Android for components
with animated `opacity`, `transform`, `backgroundColor`, `borderColor`
and `borderRadius`. If there is some other style or prop in
`useAnimatedStyle` or `useAnimatedProps`, all props of this component
will be updated using the default path (i.e. via `ShadowTree::commit`).
In some cases, this can lead to non-consistent animation behavior, e.g.
when one component will be animated using the fast path and the other
will be animated using the default path.

## Performance measurements

Initial benchmarks show 2-4x speedup on Android emulator in debug mode.

## Further improvements

* Use JNI direct access instead of copying int and float arrays
* O(1) string lookup (hashes?) in string to command conversion functions
* Make `intBuffer` and `doubleBuffer` static to avoid allocating them on
each animation frame
* Use `intBuffer[i++]` instead of `intIterator.nextInt()`
* Move serializer and parser logic to separate files

## Test plan