Update getting-started.mdx with the right installation command for re… (#7885)
…-animated 4

The next tag doesn't exist anymore when user tries to install
re-animated 4 with npm install react-native-reanimated@next or bun add
react-native-reanimated@next. This PR updates the documentation to use
the right tag

<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Sebastian Kuśnierz <47144579+skusnierz@users.noreply.github.com>