docs: Fix issue causing code blocks to appear on a single line (#7882)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixed the issue where code blocks appeared in a single line, as shown in
the image below.

<img width="973" height="483" alt="image"
src="https://github.com/user-attachments/assets/3dfdbd75-587e-461d-a008-6122cdbe4e00"
/>

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

🚀

Co-authored-by: Sebastian Kuśnierz <47144579+skusnierz@users.noreply.github.com>