feat: Custom CSS properties preprocessors for different component types (#7741)
## Summary

### Description

This PR adds a possibility to specify custom prop preprocessors on the
JS side based on the native react component name (not Fabric component
name, which is stored in the `ShadowNode`, because we cannot get this
name in JS - more details below).

This PR also reverts changes introduced in the #7734 PR. I no longer
need to use `folly::dynamic` to store keyframes object in C++ as I found
a way to distinguish components on the JS side and obtain the native
component name in there.

### Inconsistent component names

We can obtain `viewName` on the JS side which differs from the
`componentName` stored in the `ShadowNodeFamily` on Fabric. This is an
inconsistency in the react-native itself, but, fortunately, there is a
function that converts the name available in JS to the Fabric component
name `componentNameByReactViewName` available under
`react/renderer/componentregistry/componentNameByReactViewName.cpp`

<details>
<summary>Function implementation</summary>

```cpp
std::string componentNameByReactViewName(std::string viewName) {
  // We need this function only for the transition period;
  // eventually, all names will be unified.

  // TODO T97384889: unify component names between JS - Android - iOS - C++
  std::string rctPrefix("RCT");
  if (std::mismatch(rctPrefix.begin(), rctPrefix.end(), viewName.begin())
          .first == rctPrefix.end()) {
    // If `viewName` has "RCT" prefix, remove it.
    viewName.erase(0, rctPrefix.length());
  }

  // Fabric uses slightly new names for Text components because of differences
  // in semantic.
  if (viewName == "Text") {
    return "Paragraph";
  }

  // TODO T63839307: remove this condition after deleting TextInlineImage from
  // old renderer code
  if (viewName == "TextInlineImage") {
    return "Image";
  }
  if (viewName == "VirtualText") {
    return "Text";
  }

  if (viewName == "ImageView") {
    return "Image";
  }

  if (viewName == "AndroidHorizontalScrollView") {
    return "ScrollView";
  }

  if (viewName == "RKShimmeringView") {
    return "ShimmeringView";
  }

  if (viewName == "RefreshControl") {
    return "PullToRefreshView";
  }

  // We need this temporarily for testing purposes until we have proper
  // implementation of core components.
  // iOS-only
  if (viewName == "ScrollContentView") {
    return "View";
  }

  // iOS-only
  if (viewName == "MultilineTextInputView" ||
      viewName == "SinglelineTextInputView") {
    return "TextInput";
  }

  return viewName;
}
```
</summary>

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>