chore: Add cpp formatting to the pre-commit hook (#7888)
## Summary

It is annoying that the code I commit is often not properly formatted
and I have to run the formatting command manually and commit the code
again. For some reason my local cpp formatter formats code in a bit
different way than it is supposed and I don't know how to fix it yet.

Nevertheless, I feel that adding a pre-commit step that fixes formatting
in modified cpp files is beneficial and doesn't cost much.