Remove Native View (#7898)
## Summary
We currently don't use the Native View because we're still experimenting
with replacing commit hooks, etc. With @MatiPl01, we've decided to
remove those changes from the main branch for now because they are
causing some issues. We'll move them later to a branch with the
experimental implementation of `display: content`.
Related PRs:
- https://github.com/software-mansion/react-native-reanimated/pull/7366
- https://github.com/software-mansion/react-native-reanimated/pull/7380
## Test plan
See at CI