feat: Static props support in SVG CSS animations (#7783)
## Summary

This PR adds support for static props in CSS animations of SVG
components from `react-native-svg`. Basically, it adds all properties
from the `props` object (instead of the style object which is not used
by svg components).

This PR also disables smooth interpolation between `%` and `px` values
in SVG as we cannot get proper parent component dimensions from the
ShadowNode `frame`, which contains invalid values. I talked about this
with @jakex7 and we decided that we will handle this case in the same
way as keywords interpolation (i.e. the value will jump to the next one
immediately on `50%` of the interpolation progress between keyframes).
Interpolation of `px` to `px` and `%` to `%` will work properly.

## Example recording


https://github.com/user-attachments/assets/c64f2bef-d79a-4431-9974-ab02b2346db4