fix: iOS crash caused by the `IOSUIScheduler` (#7953)
## Summary

This PR aims to resolve an iOS crash happening in the `scheduleOnUI`
method of the `IOSUIScheduler` class, the occurrence of this crash is
completely occasional and it's not easy to be reproduced locally, at
least I didn't manage to do it until now, I caught this crash by our
Sentry error monitoring, and the fix in this PR is totally from logic
inference.

## Test plan

I already applied the fix in this PR in our prod app for more than one
month by patching the library, it proves to work and I didn't find any
new relevant crash occurrences in our app since this fix, so I submitted
this PR to expect we can get rid of the library patching.

---------

Co-authored-by: Zhi Zhou <zhi.zhou@getjerry.com>