fix: Small CSS issues introduced in recent PRs (#7972)
## Summary

This PR fixes 2 issues:

1. Release build
[failing](https://github.com/software-mansion/react-native-reanimated/actions/runs/16658346570/job/47149005018#step:12:3141)
because of the missing implementation of the `isValidColorString` method
- I mistakenly put it within the `#ifdef DEBUG` block
2. Error thrown during the app development on hot reloads when the new
code was added in the app - caused by the check if interpolators for the
specific component have already been registered. I just removed this
check as it is unlikely that someone registers different interpolators
for the view with the same name and I don't know if there is a better
solution to fix the hot reloads issue.