fix: CSS transition color crashes (#8002)
## Summary

This issue was introduced in the #7784 PR and was caused by the
`getDouble()` call on the `folly::dynamic` value. I assumed that all
values that come from JS are double (since this is how JS represents all
numbers) but, in fact, the value was stored as the 64-bit integer in the
`folly::dynamic` and the `.getDouble()` call caused crashes.

To fix the issue, I decided to use the safer method `.asInt`, which will
convert the number to the integer type (if it is not already an int)
instead of throwing error. I also changed the `double` constructor to
the `int64_t` constructor as the integer type is more correct than the
`double` type.

## Example recording

### Before


https://github.com/user-attachments/assets/6b5dc444-7049-4687-a36d-55dd7c9f33ac

### After


https://github.com/user-attachments/assets/d9c22957-ba62-41ab-897f-8e31ebac90c0


## Test plan

Open the `Transitions -> Test Examples -> Playground` example and press
on the **Change state** button a few times in a row.