fix: Invalid web LA keyframe processing when from/to offset is used (#8035)
## Summary

The previous implementation used an invalid type for the
`KeyframeDefinitions` type assuming that all keys in
`KeyframeDefinitions` are numbers. Because of that, it also converted
all keys to numbers via the `.map(Number)` mapping which gave `NaN`
offset values in the `offsets` array. As a result, the easing shift
didn't work properly as `NaN` aren't valid offsets so we were getting
`undefined` instead of the valid keyframe definition object.

## Example recordings

### Before


https://github.com/user-attachments/assets/1dd8933c-1811-41c7-ada1-4a87c29404c3

### After


https://github.com/user-attachments/assets/7b62f9ad-8d29-4e0f-99e4-89faf2cc19c6