[Web LA] Correctly handle `Easing.bezier` (#8049)
## Summary

### Preface

Currently using `Easing.bezier` is not supported in layout animations on
web. If you try to use it in your animation, eg.:

```jsx
<Animated.View
  entering={BounceInLeft.easing(Easing.bezier(0.25, 1, 0.25, 1))}
  ...
/>
```
you will get the following warning in console:

```
[Reanimated] Selected easing is not currently supported on web.
```

This PR adds correct handling of this easing when creating animation.


### Solution

Bezier easing is defined like so:


```tsx
function bezier(
  x1: number,
  y1: number,
  x2: number,
  y2: number
): EasingFunctionFactory {
  'worklet';
  return {
    factory: () => {
      'worklet';
      return Bezier(x1, y1, x2, y2);
    },
  };
}
```

therefore to obtain points required for bezier curve, we need to access
`factory` closure. We can do so, by checking `__closure` property:

```js
easing.factory.__closure
```

Since `factory` returns `Bezier` function, we can check if this exists
in closure. If so, we are dealing with Bezier easing.


## Test plan

Tested on Basic Layout Animation example with `Easing.bezier` added.