fix: Improve animated ref value type (#8052)
## Summary

This PR improves types used for the animated ref value. The previous
implementation required the value to be an instance of the
`React.Component`, which worked in most cases. This broke in `FlashList`
v2 because the `ref` prop doesn't extend the `React.Component` type and
instead it included just types of a few methods that can be used to
obtain the view instance.

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>