fix: Incompatible WrapperRef type in RN 0.81 (#8070)
## Summary

This PR adds just one missing possible type of the `getNativeScrollRef`
return value to the `NativeScrollRef` type union.

The `NativeMethods` type was exported from RN version before as well and
it wasn't just added in RN 0.81, so it is safe to import it and use in
the union type.

## Test plan

I tested on the branch with RN 0.81 and typecheck passed with no issues