feat(Worklets): Synchronizable (#7304)
## Summary

Implementing `Synchronizable` per its definition in the [design
document](https://github.com/software-mansion/react-native-reanimated/discussions/7264),
with slight changes.

### Synchronizable

A Synchronizable is a shared memory not tied to any JavaScript Runtime.
It can be accessed from any JavaScript Runtime and C++. The value held
by Synchronizable appears to the user as a normal JavaScript value. In
its current implementation the actual value held is of `Serializable`
type.

Synchronizable can be obtained exclusively or non-exclusively, therefore
trying to access it from multiple threads at once could lead to
blocking.

Synchronizable is an advanced concept. It can be used to implement a
semi-global state for runtimes to synchronize with. As it's not tied to
any runtime, reading and writing to it doesn't preempt any runtimes and
yields better performance compared to planned `Shareable` type.

|
![synchronizable-model](https://github.com/user-attachments/assets/a84c3c7b-d6b6-4907-b76c-c36100cf84ec)
|
|:--:| 
| *Memory model of Synchronizables, `lock` and `unlock` methods were
omitted. * |

A Synchronizable has the following methods:

- **getBlocking** - exclusively obtains the Synchronizable and returns
its value, potentially blocking if either `setDirty` or `setBlocking`
are being executed on another thread.
- **getDirty** - non-exclusively obtains the Synchronizable and returns
its value. It never blocks, even if either `setDirty` or `setBlocking`
are being executed on another thread, therefore could result in a *dirty
read*.
- **setBlocking** - exclusively obtains the Synchronizable and sets it
to the provided value, potentially blocking if either `getBlocking`,
`setBlocking` or `setDirty` are being executed on another thread.
- ~~**setDirty** - non-exclusively obtains the Synchronizable and sets
it to the provided value, potentially blocking if either `getBlocking`
or `setBlocking` are being executed on another thread. It doesn't block
when `setDirty` is executed on another thread, therefore could result in
a *lost update*.~~
I decided to hold the value of the Synchronizable as a
`std::shared_ptr<Serializable>`. Because of that it's impossible to make
`getDirty` function work since several threads could try to deconstruct
the value held by the released pointer. If the Synchronizable held a
primitive, i.e. `double` or `boolean` it wouldn't be a problem.
Therefore `setBlocking` is the only viable method to set the value with.
We might iterate on that in the future.
- **lock** - locks the Synchronizable - other threads can't call
`getBlocking` or `setBlocking` until the Synchronizable is unlocked.
Multiple calls on the same thread do nothing. Calling from another
thread while the value is locked results in blocking.
- **unlock** - unlocks the Synchronizable. Does nothing on already
unlocked Synchronizable. Works only when called from the same thread
which locked the Synchronizable.

### createSynchronizable

A function that creates a Synchronizable from a given value. It returns
a Synchronizable. Synchronizable is a JavaScript object with all the
aforementioned methods and a `jsi::NativeState` which holds the
`std::shared_ptr` to the actual C++ Synchronizable.

#### Example usage

```typescript
import { createSynchronizable, scheduleOnUI } from 'react-native-worklets';

// RN Runtime, JS thread

const synchronizable = createSynchronizable({a: 42});

scheduleOnUI(() => {
    // Note that it's still possible for `getBlocking` to be
    // executed before `setDirty` and get "{a: 42}" instead.
    const value = synchronizable.getBlocking();
    console.log(value); // {a: 24}
});

// Will most likely be executed before `getBlocking`.
synchronizable.setDirty({a: 24});
```

|
![synchronizable-example](https://github.com/user-attachments/assets/27668f1a-b0db-46d0-bd20-77067566ebac)
|
|:--:| 
| *Possible flow of the Synchronizable example usage
snippet.<br>Scheduling part was omitted.* |

## Test plan

- SynchronizableExample
- New runtime test suite

## TODO

- Docs after this PR is merged.