fix: Throw error when CSS is used on unsupported components (#8116)
## Summary

This PR fixes unpredictable behavior when CSS animation/transition was
passed to the SVG component. The behavior was unpredictable because,
when the support for SVGs was disabled, on the JS side we used the base
style builder as a fallback (the one that is used for React Native
components). As a result, some properties, such as the `color` property,
were passed to the native side where we tried to create an interpolator
for them, but, since there were no native interpolators registered in
CPP, we got a weird error message that the respective interpolator was
not found.

To make the error message more user-friendly, I decided to change this
behavior and no longer use the base style builder as a fallback for all
unsupported components. Instead, I check if the component is supported
and throw an error only if the user passed some CSS properties to the
unsupported component.

## Example recordings

When `EXPERIMENTAL_CSS_ANIMATIONS_FOR_SVG_COMPONENTS` is set to `false`.

We can see that before we were able to render SVG components when
support for them was disabled but there was no animation (which is a bit
confusing to me as no error or warning was shown, so I think that
throwing an error is better). We were able to navigate between screens
until the component that has the same prop as one of the base style
builder properties was rendered.

| Before | After |
|-|-|
| <video
src="https://github.com/user-attachments/assets/e1e469ad-db6f-4c52-9d2d-ba748badddee"
/> | <video
src="https://github.com/user-attachments/assets/6efbbea5-2970-422d-b7b1-9ef0e549b727"
/> |