Add flat list with layout animations example (#8029)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds an example that uses `FlatList` to render 200 items. Each
item is an `Animated.Text` with entering, layout and exiting animations.
Each 500 ms, the visibility of the first item is toggled, triggering
layout animations for the remaining items. The list is also wrapped
within `LayoutAnimationConfig` component with `skipEntering` enabled to
prevent entering animation during first render when the screen is being
opened.


https://github.com/user-attachments/assets/2a9641de-b990-4e27-a300-2750783b1609

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->