chore: Remove unnecessary checks from css value constructors (#8185)
## Summary

This PR removes unnecessary correctness checks from CSSValues
constructors as they were just unnecessarily doing the same work as the
`canConstruct` methods do. In the end, we ended up checking if the
`folly::dynamic` or the `jsi::Value` is correct twice because these
`folly::dynamic` and `jsi::Value` constructors are never used directly
and the only place in which they are used to construct the value is the
`CSSValueVariant` class, which always calls `canConstruct` method before
calling the actual constructor.

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>