chore: Reduce CSS ValueInterpolator templating (#8198)
## Summary

This PR extracts what can be extracted from the template CSS
`ValueInterpolator` class into a base, non-template class. It also moves
the template part in to separate file and renames it to
`SimpleValueInterpolator` to match the naming convention of the CSS
values (`CSSSimpleValue` -> `SimpleValueInterpolator`,
`CSSResolvableValue` -> `ResolvableValueInterpolator`).

After these changes, I could move much of the interpolator
implementation code into a separate `cpp` file from the header file,
which will make debugging easier as we will be able to set breakpoints
in this file (we can't do this in header files).