chore(Worklets): add scope recrawling during closure creation in Babel plugin (#8204)
## Summary

During closure creation of worklets we might get into a situation where
another plugin has added an identifier but didn't update the scope.
Since the algorithm assumes that the scope is the source of truth, we
forcefully recrawl the scope once when we fail to find a binding, to fix
other plugins' mistakes.

## Test plan

Reported by @kitten, he confirmed that it fixes the problem.