fix: Initialize logger before all other things (#8227)
## Summary

The current initialization pipeline is quite tricky and it turned out
that the logger could have been used before the logger config was
written into the `global` object.

I suspect that this may be caused by the imports chain starting from the
`Animated.ts` file that imports many other files, which in the end have
imports from the
`packages/react-native-reanimated/src/animation/util.ts` file. This file
contains the following global if statement:

```ts
if (__DEV__ && ReducedMotionManager.jsValue) {
  logger.warn(
    `Reduced motion setting is enabled on this device. This warning is visible only in the development mode. Some animations will be disabled by default. You can override the behavior for individual animations, see https://docs.swmansion.com/react-native-reanimated/docs/guides/troubleshooting#reduced-motion-setting-is-enabled-on-this-device.`
  );
}
```

Because of that, this log could be executed before the
`initializeReanimatedModule` function in the `src/index.ts` file was
called, which resulted in the crash reported in this issue: #8199