chore: Move interpolator configs to a single implementation file (#8258)
## Summary

This PR moves all interpolator configs to a single file. As a result,
only a single TU (translation unit) is being created during the
compilation step when the
`packages/react-native-reanimated/Common/cpp/reanimated/CSS/InterpolatorRegistry.cpp`
includes the `#include
<reanimated/CSS/interpolation/InterpolatorFactory.h>` header that
contains multiple template functions and classes. Thanks to this change,
the implementation for templates declared in the `InterpolatorFactory.h`
will be generated just once instead of multiple times for each config
file separately as it did before.