[LA] Add check if view is mounted on Android  (#8083)
This is a temporary fix for #7493, which should ultimately be solved by
the pull model being implemented on android. The solution for now would
be to check if a view is mounted whenever we try to update an animation
frame via layout animations. Maybe we could change that so we only check
until we know the view is mounted, but for safety let's have it this way
for now.

On my end (on an emulator) it takes on average 450 μs to execute this
function for 500 views and 100 μs for 100 views. Usually, there will be
far less views animated via layout animations, so we should be fine with
this workaround for now.

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>