fix: no Babel plugin support (#8241)
## Summary

Fixes #8179.

Since we don't usually test no Worklets Babel Plugin configuration I
didn't notice that I introduced a regression that always throws when the
plugin is not detected. This PR enables this check only in web
environments.

I'm also amending the docs with the information that we discourage
opting out of the plugin since we don't test that configuration.

## Test plan

Run web example without the Babel plugin