chore: Bump react-native-pager-view to v7 (#8278)
## Summary

I was facing
[this](https://github.com/callstack/react-native-pager-view/issues/979)
issue multiple times which was very annoying while working on native
changes in reanimated. Usually, the only thing that worked was a full
cleanup of the project and reinstallation of all dependencies and pods.

I hope that the issue is fixed in v7. At least, I think we can give it a
try and see if it works after the upgrade.

I tested if the pager example in our example app works so I think it is
safe to bump the version to v7.