chore: use automatic jsx runtime and bump react-native-builder-bob (#7451)
## Summary

Fixes #7969.

In #7266 I made it so JSX is preserved to address some issues of
NativeWind but this wasn't the best solution since now we have JSX code
in .js files which is problematic in some Web environments (Jest,
Tanstack). Instead we opt-in for automatic JSX runtime. I falsely
claimed in #7266 that automatic JSX runtime was enabled.

These are effective changes:

```tsx
// source .tsx file
'use strict';

import { AnimatedView } from './View';

export function Foo() {
  return <AnimatedView />;
}
```

Change in the emitted file:

```diff
--- Foo.old.js
+++ Foo.new.js
@@ -1,7 +1,8 @@
 'use strict';
 
-import { AnimatedView } from "./View.js";
+import { AnimatedView } from './View';
+import { jsx as _jsx } from "react/jsx-runtime";
 export function Foo() {
-  return <AnimatedView />;
+  return /*#__PURE__*/_jsx(AnimatedView, {});
 }
 //# sourceMappingURL=Foo.js.map
```

Key changes are:
- JSX becomes `_jsx` with added import of `react/jsx-runtime` - I
believe you can configure the runtime on the bundler side so it's
supposed to be as general as pure JSX.
- Relative imports lose the file extension. This shouldn't introduce any
regressions - it should actually even improve the behavior for
resolution of `.web.js` files in some environments.

## Test plan

I asked @blazejkustra to check it in Expensify (Jest) and @marklawlor in
NativeWind. I also confirmed that original issue from #7969 is solved
with this change. Waiting for results 🙏.

A version of Reanimated released with these changes is available as on
npm as `react-native-reanimated@4.1-jsx-fix` for testing.