chore(Worklets): fix PlatformChecker init (#8300)
## Summary

Fixes #8285.

When initializing "PlatformChecker" in lazy import environments
`globalThis.__RUNTIME_KIND` will be executed before it's defined in
`runtimeKind` file. By changing it to function invocation we enforce
that it's properly initialized.

Also setting serialized values manually for `runOnUI` and `runOnUIAsync`
has to be done only when not in web implementation.

## Test plan

See fixed issue.