fix: Don't pass uninitialized ref to dispatchCommand, add warnings (#8327)
## Summary

This PR adds warnings shown when the `scrollTo` or the `dispatchCommand`
function is called with an uninitialized ref. It also early returns from
the `dispatchCommand` function when the ref is uninitialized (has `null`
value) and doesn't pass it to the native `dispatchCommand` method call.

I added this PR as I feel that the #8245 issue is caused by the invalid
use of the animated ref.