fix: Animated styles not registered after unfreeze (#8325)
## Summary

This PR fixes animated styles registering on freeze/unfreeze. The issue
was caused by the cleanup made in the `componentWillUnmount`, which
removes the current view tag from the animated style's view descriptors.
This view tag was never added later on because only the style object
references were compared and, because the style prop didn't change
between freeze/unfreeze, we assumed that there is no need to register
the animated styles again. This assumption was correct as long as the
`componentWillUnmount` was executed before the component was actually
unmounted (but, it turns out that this method is also called when the
component is frozen, not unmounted).

## Example recordings

The pink rectangle with the animated style attached didn't animate after
navigating back to the previously frozen screen.

| Before | After |
|-|-|
| <video
src="https://github.com/user-attachments/assets/eaaa21ae-266a-4649-8fb3-b492e80fc186"
/> | <video
src="https://github.com/user-attachments/assets/744eb171-d231-49aa-92f2-e89174dd2b15"
/> |