docs: Incorporate docs into monorepo (#8304)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

More well-organized and sorted version of #8298 

This PR adds docs as a part of monorepo ecosystem, and makes them use
yarn v4.

To-do:

- [x] fix ymls
- [x] check if docs publish CI needs to be updated - probably not as it
worked locally, but needs to make sure react-native-reanimated is not
needed to be build beforehand
- [x] precommit
- [x] linting CIs

## Test plan

Run `yarn start` in both monorepos. **IMPORTANT!** Also run `yarn build
&& yarn serve` in reanimated - it won't work in worklets as worklet
documentation is done using some special under the table tricks to make
it possible for two docs to work in one monorepo. Take a look at CIs
changes, as well as new build path of docs.