chore: Return unique_ptr from TransformOperationInterpolator interpolate method (#8342)
## Summary

As the title states, this PR changes the return type of the
`interpolate` method declared on the `TransformOperationInterpolator`
class to the `std::unique_ptr` from `folly::dynamic`. This is necessary
to add support for SVG transforms handling were we have to create a
single matrix from interpolated operation and can't convert transforms
array to `folly::dynamic::array` like we do in plain transforms.