fix(Worklets): Bundle Mode accidentally toggling in Expo54 (#8335)
## Summary

Expo 54 performs eager imports contrary to Metro which is lazy. This
results in `workletRuntimeEntry.ts` file to be executed and for the
`globalThis._WORKLETS_BUNDLE_MODE` flag to be set wrongfully to `true`.

To handle both eager and lazy imports environments I decided to handle
the setting of the flag's value in Worklets Babel Plugin. Worklets Babel
Plugin makes it so the expression in `workletRuntimeEntry.ts`:

```ts
globalThis._WORKLETS_BUNDLE_MODE = false
```

becomes

```ts
globalThis._WORKLETS_BUNDLE_MODE = true
``` 

when Bundle Mode is enabled.

## Test plan

I tested these changes with an Expo 54 app and they worked.