docs: Remove 0.82 from supported versions in v3 (#8374)
## Summary

Starting from React Native `0.82`, the New Architecture can no longer be
disabled and the Old Architecture (Paper) is not supported at all (see
this [PR](https://github.com/facebook/react-native/pull/53026) that is a
part of the `0.82` release).

Because of that, I decided to remove the `0.82` support from the
`compatibility.json`. I think we will have to stop the development of
the v3 on the RN `0.81` version as I see no point in adding v3 support
for RN `0.82` on Fabric, while leaving Paper not supported.