fix: scrollTo is not a function on web (#8362)
## Summary

This PR inverts the `if` condition that I forgot to invert when applying
a
[cherry-pick](https://github.com/software-mansion/react-native-reanimated/pull/8334/files#diff-a98b820e995e73c1ce7fe27c3ad58e503b5361258bea29c41eafdcb95fdf1a09)
before the `3.19.2` release and fixes the #8361 issue.

## Example recordings

### Before

#### Uninitialized ref

Should show the warning shown in the next example instead of throwing an
error.


https://github.com/user-attachments/assets/9f44c11e-cd57-478f-a999-e30b2ad2d451

#### Initialized ref

Should just scroll properly instead of showing the warning.


https://github.com/user-attachments/assets/2da5a041-d960-46f9-82ae-14a4a4db2a48

### After

#### Uninitialized ref

Fine, properly shows a warning if a ref is not initialized.


https://github.com/user-attachments/assets/db8a2ff6-6001-4443-a927-4bd91be573eb

#### Initialized ref

Fine, properly scrolls when the ref is initialized.


https://github.com/user-attachments/assets/bc8ac103-ac3c-4787-bbd1-5592e82e9e2f